requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
pypdf==6.0.0
numpy==2.2.2
faiss-cpu==1.10.0
sentence-transformers==3.3.1
langdetect==1.0.9
